+++ src/pages/api/auth/forgot-password/reset.ts
@@ -1,8 +1,8 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { getPasswordResetTokenWithToken } from "../../../../utils/passwordResetTokens";
+import { getPasswordResetTokenWithToken } from "../../../../database/utils/passwordResetTokens";
 import bcrypt from 'bcryptjs'
 import { updateUserPassword } from "../../../../utils/users";
-import {deletePasswordResetToken} from '../../../../utils/passwordResetTokens'
+import {deletePasswordResetToken} from '../../../../database/utils/passwordResetTokens'
 
 export default async function ResetPassword(req:NextApiRequest, res:NextApiResponse) {
